Django/requirements.txt
appdata==2.2.1
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
colored==2.3.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
docopt==0.6.2
dotenv==0.9.9
idna==3.10
pillow==11.3.0
pipreqs==0.4.13
psycopg2==2.9.10
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
yarg==0.1.10
